+++ R/import_ghcn_stations.R
@@ -27,7 +27,7 @@
 #'   other.
 #'
 #' @param provider When `return = "map"`, by default a map will be created in
-#'   which readers may toggle between a vector base map and a satellite/aerial
+#'   which readers may toggle between a vector street map and a satellite/aerial
 #'   image. `provider` allows users to override this default; see
 #'   \url{http://leaflet-extras.github.io/leaflet-providers/preview/} for a list
 #'   of all base maps that can be used. Base maps can be toggled using a layer
@@ -64,7 +64,10 @@ import_ghcn_stations <-
     lng = NULL,
     crs = 4326,
     n_max = 10L,
-    provider = c("OSM" = "OpenStreetMap", "Satellite" = "Esri.WorldImagery"),
+    provider = c(
+      "Street Map" = "CartoDB.Voyager",
+      "Satellite" = "Esri.WorldImagery"
+    ),
     database = c("hourly", "daily"),
     return = c("table", "sf", "map")
   ) {