+++ R/import_isd_stations.R
@@ -50,7 +50,10 @@ import_isd_stations <- function(
   crs = 4326,
   n_max = 10,
   end_year = "current",
-  provider = c("OSM" = "OpenStreetMap", "Satellite" = "Esri.WorldImagery"),
+  provider = c(
+    "Street Map" = "CartoDB.Voyager",
+    "Satellite" = "Esri.WorldImagery"
+  ),
   return = c("table", "sf", "map")
 ) {
   ## read the meta data